Fix prettier problems related to linting json files Add shared types to maintenance-app Use shared types on maintenance-app/backend Start adding controls for setup & reset in maintenance-app/frontend
