+++ packages/maintenance-app/types/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@maintenance-app/types",
+  "version": "0.0.0",
+  "description": "Veto app shared types",
+  "author": "Veriphi",
+  "license": "PRIVATE",
+  "private": true,
+  "main": "out/index.js",
+  "type": "out/index.d.ts",
+  "scripts": {
+    "postinstall": "tsc -p . --outDir out",
+    "start": "tsc -p . --outDir out --watch",
+    "build": "tsc -p . --outDir out"
+  }
+}
