+++ packages/veto-app/frontend/package.json
@@ -3,32 +3,6 @@
   "version": "0.0.0",
   "description": "Veto front-end application",
   "private": true,
-  "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.27",
-    "@fortawesome/free-solid-svg-icons": "^5.12.1",
-    "@fortawesome/react-fontawesome": "^0.1.8",
-    "@reach/router": "^1.2.1",
-    "@veriphi/veto-ui": "^0.2.0",
-    "@veto/config": "0.0.0",
-    "axios": "^0.19.2",
-    "emotion-theming": "^10.0.27",
-    "gun": "^0.2020.116",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-scripts": "^3.4.0",
-    "swr": "^0.2.0",
-    "typescript": "^3.7.4"
-  },
-  "devDependencies": {
-    "@testing-library/jest-dom": "^5.5.0",
-    "@testing-library/react": "^10.0.2",
-    "@testing-library/user-event": "^10.0.1",
-    "@types/jest": "^25.1.3",
-    "@types/node": "^12.0.0",
-    "@types/react": "^16.9.0",
-    "@types/react-dom": "^16.9.0",
-    "ts-jest": "^25.2.1"
-  },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
@@ -50,5 +24,32 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^1.2.27",
+    "@fortawesome/free-solid-svg-icons": "^5.12.1",
+    "@fortawesome/react-fontawesome": "^0.1.8",
+    "@reach/router": "^1.2.1",
+    "@rehooks/local-storage": "^2.1.1",
+    "@veriphi/veto-ui": "^0.2.0",
+    "@veto/config": "0.0.0",
+    "axios": "^0.19.2",
+    "emotion-theming": "^10.0.27",
+    "gun": "^0.2020.116",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "react-scripts": "^3.4.0",
+    "swr": "^0.2.0",
+    "typescript": "^3.7.4"
+  },
+  "devDependencies": {
+    "@testing-library/jest-dom": "^5.5.0",
+    "@testing-library/react": "^10.0.2",
+    "@testing-library/user-event": "^10.0.1",
+    "@types/jest": "^25.1.3",
+    "@types/node": "^12.0.0",
+    "@types/react": "^16.9.0",
+    "@types/react-dom": "^16.9.0",
+    "ts-jest": "^25.2.1"
   }
 }
